Started saving trees to database
While testing, i discovered a bug. You can see the effects of the bug by uncommenting line 56 of RoomPomodoro.jsx, where it prints a message about trees to the console. Every time you start the timer again after a break, there are more messages than the time before. Is this because there are more timers, more callbacks to the same time, or is the useEffect being called more due to a change in dependencies? further research necessary. Also, additional types of trees are untested.