prototypes for treeDisplayBanner and treeDisplaySelector
treeDisplayBanner is a very simple compenent that displays 5 trees from an array of treeIDs, each taking up 1/10th of their container. treeDisplaySelector ins unfinished, but when finished, it will have 5 dropdown menus that allow users to pick which trees to select for their display banner. It will not allow users to pick trees they do not have, and will not allow them to exceed the current inventory of any given tree.