updated new component prototypes, solved timer bug
the bug with the timer was "solved" by placing the code that updates the database when a new tree is grown in the TimerBar component rather than the RoomPomodoro component.

The treeDisplaySelector component is almost mature. It just needs to send the selected trees to the database.

The treeDisplayBanner component is also almost mature. It just needs to load the array of selected trees from the database instead of using a toy example.